+++ src/App.tsx
@@ -7,12 +7,10 @@ function App() {
 			<header>
 				<nav>
 					<Link to="/">Accueil</Link>
-					<Link to="/my_profile">Mon profil</Link>
-					<Link to="/discover_profiles">Découvrir des profils</Link>
+					<Link to="/myProfile">Mon profil</Link>
+					<Link to="/profiles">Découvrir des profils</Link>
 				</nav>
 			</header>
-			<h1>coucou</h1>
-			<h2>toto</h2>
 			<Outlet />
 		</>
 	);
