--- src/pages/Discover_profiles.tsx
@@ -1,6 +1,0 @@
-import "../css/Discover_profiles.css";
-
-function Discover_profiles() {
-	return <p>Découvrir des profils</p>;
-}
-export default Discover_profiles;
