--- src/pages/My_profile.tsx
@@ -1,6 +1,0 @@
-import "../css/My_profile.css";
-
-function My_profile() {
-	return <p>Mon profil</p>;
-}
-export default My_profile;
