+++ src/pages/myProfile/MyProfile.tsx
@@ -1,0 +1,6 @@
+import "./MyProfile.css";
+
+function MyProfile() {
+	return <p>Mon profil</p>;
+}
+export default MyProfile;
