+++ src/pages/profiles/Profiles.tsx
@@ -1,0 +1,6 @@
+import "./Profiles.css";
+
+function Profiles() {
+	return <p>Découvrir des profils</p>;
+}
+export default Profiles;
